Implement unbounded iterations #69
The switch -maxiters recognizes a new argument, 'Infinity' to indicate
that the number of iteration should not be bounded.
If a number is given, it has to be between zero and Integer.MAX_VALUE.

The implementation stores these values as doubles. This allows to stick
to comparisons like if(iters < max_iters) as +Infinity compares greater to
all integer values.

+++ prism/include/jni/prism_PrismNative.h
@@ -66,10 +66,10 @@ JNIEXPORT void JNICALL Java_prism_PrismNative_PN_1SetTermCritParam
 /*
  * Class:     prism_PrismNative
  * Method:    PN_SetMaxIters
- * Signature: (I)V
+ * Signature: (D)V
  */
 JNIEXPORT void JNICALL Java_prism_PrismNative_PN_1SetMaxIters
-  (JNIEnv *, jclass, jint);
+  (JNIEnv *, jclass, jdouble);
 
 /*
  * Class:     prism_PrismNative
